+++ client/src/Redux/actions/authActions.js
@@ -11,7 +11,7 @@ export const clearUser = () => ({
 export const registerUser = (userData, redirect) => dispatch =>
 	axios
 		.post('/api/users/register', userData)
-		.then(() => redirect.push('/login'))
+		.then(() => redirect.push('/'))
 		.catch(err => {
 			dispatch({
 				type: types.GET_ERRORS,
