+++ client/src/components/App.js
@@ -4,8 +4,6 @@ import { BrowserRouter as Router, Switch } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import { logoutUser } from '../Redux/actions/authActions';
-import Login from './Login/Login';
-import Register from './Register/Register';
 import Landing from './Landing/Landing';
 import PrivateRoute from './Auth/PrivateRoute';
 import PublicRoute from './Auth/PublicRoute';
@@ -21,8 +19,6 @@ const App = props => {
 				<Nav isAuthenticated={props.isAuthenticated} logout={props.logout} />
 				<Switch>
 					<PublicRoute exact path="/" component={Landing} isAuthenticated={props.isAuthenticated} />
-					<PublicRoute path="/register" component={Register} isAuthenticated={props.isAuthenticated} />
-					<PublicRoute path="/login" component={Login} isAuthenticated={props.isAuthenticated} />
 					<PrivateRoute path="/dashboard" component={Dashboard} isAuthenticated={props.isAuthenticated} />
 					<PrivateRoute path="/journal" component={Journal} isAuthenticated={props.isAuthenticated} />
 				</Switch>
