--- client/src/components/common/Input.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Input = () => {
-	return (
-		<div className="input__box">
-			<input type="text" name="input-field" />
-		</div>
-	);
-};
-
-export default Input;
