--- client/src/components/common/InputButton.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const InputButton = props => {
-	return (
-		<div>
-			<button>Button</button>
-		</div>
-	);
-};
-
-InputButton.propTypes = {};
-
-export default InputButton;
