+++ client/src/styles/_button.scss
@@ -62,7 +62,7 @@
 
 	&--primary {
 		background: var(--accent-color);
-		color: var(--font-color-mid);
+		color: var(--primary-light);
 
 		&:hover,
 		&:visited {
